add redux persist storage
