+++ package.json
@@ -18,6 +18,7 @@
     "react-router-dom": "^7.3.0",
     "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
+    "redux-persist": "^6.0.0",
     "reselect": "^5.1.1",
     "styled-components": "^6.1.15"
   },
