+++ src/store/categories.ts
